Refactored auth controllers
